refactor: move segment types
