+++ crates/lib/src/dialects/ansi.rs
@@ -1,16 +1,10 @@
-use std::fmt::Debug;
 use std::sync::Arc;
 
-use itertools::{enumerate, Itertools};
-use smol_str::{SmolStr, ToSmolStr};
+use itertools::Itertools;
 
 use super::ansi_keywords::{ANSI_RESERVED_KEYWORDS, ANSI_UNRESERVED_KEYWORDS};
-use super::{SyntaxKind, SyntaxSet};
+use super::SyntaxKind;
 use crate::core::dialects::base::Dialect;
-use crate::core::dialects::common::{AliasInfo, ColumnAliasInfo};
-use crate::core::dialects::init::DialectKind;
-use crate::core::errors::SQLParseError;
-use crate::core::parser::context::ParseContext;
 use crate::core::parser::grammar::anyof::{one_of, optionally_bracketed, AnyNumberOf};
 use crate::core::parser::grammar::base::{Anything, Nothing, Ref};
 use crate::core::parser::grammar::conditional::Conditional;
@@ -20,7 +14,6 @@ use crate::core::parser::lexer::{Matcher, Pattern};
 use crate::core::parser::matchable::Matchable;
 use crate::core::parser::node_matcher::NodeMatcher;
 use crate::core::parser::parsers::{MultiStringParser, RegexParser, StringParser, TypedParser};
-use crate::core::parser::segments::base::{ErasedSegment, SegmentBuilder, Tables};
 use crate::core::parser::segments::bracketed::BracketedSegmentMatcher;
 use crate::core::parser::segments::generator::SegmentGenerator;
 use crate::core::parser::segments::meta::MetaSegment;
@@ -40,6 +33,7 @@ impl<T> BoxedE for T {
         Arc::new(self)
     }
 }
+
 pub fn dialect() -> Dialect {
     raw_dialect().config(|this| this.expand())
 }
@@ -4809,523 +4803,6 @@ pub fn statement_segment() -> Arc<dyn Matchable> {
 
 pub fn wildcard_expression_segment() -> Arc<dyn Matchable> {
     Sequence::new(vec![Ref::new("WildcardIdentifierSegment").boxed()]).to_matchable()
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct FileSegment;
-impl FileSegment {
-    pub fn of(
-        tables: &Tables,
-        dialect: DialectKind,
-        segments: Vec<ErasedSegment>,
-    ) -> ErasedSegment {
-        SegmentBuilder::node(tables.next_id(), SyntaxKind::File, dialect, segments)
-            .position_from_segments()
-            .finish()
-    }
-
-    pub fn root_parse(
-        &self,
-        tables: &Tables,
-        dialect: DialectKind,
-        segments: &[ErasedSegment],
-        parse_context: &mut ParseContext,
-        _f_name: Option<String>,
-    ) -> Result<ErasedSegment, SQLParseError> {
-        let start_idx = segments.iter().position(|segment| segment.is_code()).unwrap_or(0) as u32;
-
-        let end_idx = segments
-            .iter()
-            .rposition(|segment| segment.is_code())
-            .map_or(start_idx, |idx| idx as u32 + 1);
-
-        if start_idx == end_idx {
-            return Ok(FileSegment::of(tables, dialect, segments.to_vec()));
-        }
-
-        let final_seg = segments.last().unwrap();
-        assert!(final_seg.get_position_marker().is_some());
-
-        let file_segment = parse_context.dialect().r#ref("FileSegment");
-
-        let match_result = file_segment.match_grammar().unwrap().match_segments(
-            &segments[..end_idx as usize],
-            start_idx,
-            parse_context,
-        )?;
-
-        let match_span = match_result.span;
-        let has_match = match_result.has_match();
-        let mut matched = match_result.apply(tables, dialect, segments);
-        let unmatched = &segments[match_span.end as usize..end_idx as usize];
-
-        let content: &[ErasedSegment] = if !has_match {
-            &[SegmentBuilder::node(
-                tables.next_id(),
-                SyntaxKind::Unparsable,
-                dialect,
-                segments[start_idx as usize..end_idx as usize].to_vec(),
-            )
-            .position_from_segments()
-            .finish()]
-        } else if !unmatched.is_empty() {
-            let idx = unmatched.iter().position(|it| it.is_code()).unwrap_or(unmatched.len());
-            let (head, tail) = unmatched.split_at(idx);
-
-            matched.extend_from_slice(head);
-            matched.push(
-                SegmentBuilder::node(tables.next_id(), SyntaxKind::File, dialect, tail.to_vec())
-                    .position_from_segments()
-                    .finish(),
-            );
-            &matched
-        } else {
-            matched.extend_from_slice(unmatched);
-            &matched
-        };
-
-        Ok(Self::of(
-            tables,
-            dialect,
-            [&segments[..start_idx as usize], content, &segments[end_idx as usize..]].concat(),
-        ))
-    }
-}
-
-pub struct FromExpressionElementSegment(pub ErasedSegment);
-pub struct FromClauseSegment(pub ErasedSegment);
-
-impl FromClauseSegment {
-    pub fn eventual_aliases(&self) -> Vec<(ErasedSegment, AliasInfo)> {
-        let mut buff = Vec::new();
-        let mut direct_table_children = Vec::new();
-        let mut join_clauses = Vec::new();
-
-        for from_expression in
-            self.0.children(const { &SyntaxSet::new(&[SyntaxKind::FromExpression]) })
-        {
-            direct_table_children.extend(
-                from_expression
-                    .children(const { &SyntaxSet::new(&[SyntaxKind::FromExpressionElement]) }),
-            );
-            join_clauses.extend(
-                from_expression.children(const { &SyntaxSet::new(&[SyntaxKind::JoinClause]) }),
-            );
-        }
-
-        for clause in &direct_table_children {
-            let tmp;
-
-            let alias = FromExpressionElementSegment(clause.clone()).eventual_alias();
-
-            let table_expr = if direct_table_children.contains(clause) {
-                clause
-            } else {
-                tmp = clause
-                    .child(const { &SyntaxSet::new(&[SyntaxKind::FromExpressionElement]) })
-                    .unwrap();
-                &tmp
-            };
-
-            buff.push((table_expr.clone(), alias));
-        }
-
-        for clause in join_clauses {
-            let aliases = JoinClauseSegment(clause).eventual_aliases();
-
-            if !aliases.is_empty() {
-                buff.extend(aliases);
-            }
-        }
-
-        buff
-    }
-}
-
-#[derive(Clone)]
-pub struct SelectClauseElementSegment(pub ErasedSegment);
-
-impl SelectClauseElementSegment {
-    pub fn alias(&self) -> Option<ColumnAliasInfo> {
-        let alias_expression_segment = self
-            .0
-            .recursive_crawl(
-                const { &SyntaxSet::new(&[SyntaxKind::AliasExpression]) },
-                true,
-                &SyntaxSet::EMPTY,
-                true,
-            )
-            .first()?
-            .clone();
-
-        let alias_identifier_segment = alias_expression_segment.segments().iter().find(|it| {
-            matches!(it.get_type(), SyntaxKind::NakedIdentifier | SyntaxKind::Identifier)
-        })?;
-
-        let aliased_segment = self
-            .0
-            .segments()
-            .iter()
-            .find(|&s| !s.is_whitespace() && !s.is_meta() && s != &alias_expression_segment)
-            .unwrap();
-
-        let mut column_reference_segments = Vec::new();
-        if aliased_segment.is_type(SyntaxKind::ColumnReference) {
-            column_reference_segments.push(aliased_segment.clone());
-        } else {
-            column_reference_segments.extend(aliased_segment.recursive_crawl(
-                const { &SyntaxSet::new(&[SyntaxKind::ColumnReference]) },
-                true,
-                &SyntaxSet::EMPTY,
-                true,
-            ));
-        }
-
-        Some(ColumnAliasInfo {
-            alias_identifier_name: alias_identifier_segment.raw().into(),
-            aliased_segment: aliased_segment.clone(),
-            column_reference_segments,
-        })
-    }
-}
-
-impl FromExpressionElementSegment {
-    pub fn eventual_alias(&self) -> AliasInfo {
-        let mut tbl_expression =
-            self.0.child(const { &SyntaxSet::new(&[SyntaxKind::TableExpression]) }).or_else(|| {
-                self.0.child(const { &SyntaxSet::new(&[SyntaxKind::Bracketed]) }).and_then(
-                    |bracketed| {
-                        bracketed.child(const { &SyntaxSet::new(&[SyntaxKind::TableExpression]) })
-                    },
-                )
-            });
-
-        if let Some(tbl_expression_inner) = &tbl_expression
-            && tbl_expression_inner
-                .child(const { &SyntaxSet::new(&[SyntaxKind::ObjectReference, SyntaxKind::TableReference]) })
-                .is_none()
-        {
-            let bracketed = tbl_expression_inner.child(const { &SyntaxSet::new(&[SyntaxKind::Bracketed]) });
-            if let Some(bracketed) = bracketed {
-                tbl_expression = bracketed.child(const { &SyntaxSet::new(&[SyntaxKind::TableExpression]) });
-            }
-        }
-
-        let reference = tbl_expression.and_then(|tbl_expression| {
-            tbl_expression.child(const { &SyntaxSet::new(&[SyntaxKind::ObjectReference, SyntaxKind::TableReference]) })
-        });
-
-        let reference = reference.as_ref().map(|reference| reference.reference());
-
-        let alias_expression =
-            self.0.child(const { &SyntaxSet::new(&[SyntaxKind::AliasExpression]) });
-        if let Some(alias_expression) = alias_expression {
-            let segment = alias_expression.child(
-                const { &SyntaxSet::new(&[SyntaxKind::Identifier, SyntaxKind::NakedIdentifier]) },
-            );
-            if let Some(segment) = segment {
-                return AliasInfo {
-                    ref_str: segment.raw().into(),
-                    segment: segment.into(),
-                    aliased: true,
-                    from_expression_element: self.0.clone(),
-                    alias_expression: alias_expression.into(),
-                    object_reference: reference.map(|it| it.clone().0),
-                };
-            }
-        }
-
-        if let Some(reference) = &reference {
-            let references = reference.iter_raw_references();
-
-            if !references.is_empty() {
-                let penultimate_ref = references.last().unwrap();
-                return AliasInfo {
-                    ref_str: penultimate_ref.part.clone().into(),
-                    segment: penultimate_ref.segments[0].clone().into(),
-                    aliased: false,
-                    from_expression_element: self.0.clone(),
-                    alias_expression: None,
-                    object_reference: reference.clone().0.into(),
-                };
-            }
-        }
-
-        AliasInfo {
-            ref_str: SmolStr::new_static(""),
-            segment: None,
-            aliased: false,
-            from_expression_element: self.0.clone(),
-            alias_expression: None,
-            object_reference: reference.map(|it| it.clone().0),
-        }
-    }
-}
-
-#[derive(PartialEq, Eq, PartialOrd, Ord, Clone, Copy)]
-pub enum ObjectReferenceLevel {
-    Object = 1,
-    Table = 2,
-    Schema = 3,
-}
-
-#[derive(Clone, Debug)]
-pub struct ObjectReferencePart {
-    pub part: String,
-    pub segments: Vec<ErasedSegment>,
-}
-
-#[derive(Clone)]
-pub struct ObjectReferenceSegment(pub ErasedSegment, pub ObjectReferenceKind);
-
-#[derive(Clone)]
-pub enum ObjectReferenceKind {
-    Object,
-    Table,
-    WildcardIdentifier,
-}
-
-impl ObjectReferenceSegment {
-    pub fn is_qualified(&self) -> bool {
-        self.iter_raw_references().len() > 1
-    }
-
-    pub fn qualification(&self) -> &'static str {
-        if self.is_qualified() { "qualified" } else { "unqualified" }
-    }
-
-    pub fn extract_possible_references(
-        &self,
-        level: ObjectReferenceLevel,
-        dialect: DialectKind,
-    ) -> Vec<ObjectReferencePart> {
-        let refs = self.iter_raw_references();
-
-        match dialect {
-            DialectKind::Bigquery => {
-                if level == ObjectReferenceLevel::Schema && refs.len() >= 3 {
-                    return vec![refs[0].clone()];
-                }
-
-                if level == ObjectReferenceLevel::Table {
-                    return refs.into_iter().take(3).collect_vec();
-                }
-
-                if level == ObjectReferenceLevel::Object && refs.len() >= 3 {
-                    return vec![refs[1].clone(), refs[2].clone()];
-                }
-
-                self.extract_possible_references(level, DialectKind::Ansi)
-            }
-            _ => {
-                let level = level as usize;
-                if refs.len() >= level && level > 0 {
-                    refs.get(refs.len() - level).cloned().into_iter().collect()
-                } else {
-                    vec![]
-                }
-            }
-        }
-    }
-
-    pub fn extract_possible_multipart_references(
-        &self,
-        levels: &[ObjectReferenceLevel],
-    ) -> Vec<Vec<ObjectReferencePart>> {
-        self.extract_possible_multipart_references_inner(levels, self.0.dialect())
-    }
-
-    pub fn extract_possible_multipart_references_inner(
-        &self,
-        levels: &[ObjectReferenceLevel],
-        dialect_kind: DialectKind,
-    ) -> Vec<Vec<ObjectReferencePart>> {
-        match dialect_kind {
-            DialectKind::Bigquery => {
-                let levels_tmp: Vec<_> = levels.iter().map(|level| *level as usize).collect();
-                let min_level: usize = *levels_tmp.iter().min().unwrap();
-                let max_level: usize = *levels_tmp.iter().max().unwrap();
-                let refs = self.iter_raw_references();
-
-                if max_level == ObjectReferenceLevel::Schema as usize && refs.len() >= 3 {
-                    return vec![(refs[0..=max_level - min_level].to_vec())];
-                }
-
-                self.extract_possible_multipart_references_inner(levels, DialectKind::Ansi)
-            }
-            _ => {
-                let refs = self.iter_raw_references();
-                let mut sorted_levels = levels.to_vec();
-                sorted_levels.sort_unstable();
-
-                if let (Some(&min_level), Some(&max_level)) =
-                    (sorted_levels.first(), sorted_levels.last())
-                {
-                    if refs.len() >= max_level as usize {
-                        let start = refs.len() - max_level as usize;
-                        let end = refs.len() - min_level as usize + 1;
-                        if start < end {
-                            return vec![refs[start..end].to_vec()];
-                        }
-                    }
-                }
-                vec![]
-            }
-        }
-    }
-    pub fn iter_raw_references(&self) -> Vec<ObjectReferencePart> {
-        match self.1 {
-            ObjectReferenceKind::Table if self.0.dialect() == DialectKind::Bigquery => {
-                let mut acc = Vec::new();
-                let mut parts = Vec::new();
-                let mut elems_for_parts = Vec::new();
-
-                let mut flush =
-                    |parts: &mut Vec<SmolStr>, elems_for_parts: &mut Vec<ErasedSegment>| {
-                        acc.push(ObjectReferencePart {
-                            part: std::mem::take(parts).iter().join(""),
-                            segments: std::mem::take(elems_for_parts),
-                        });
-                    };
-
-                for elem in self.0.recursive_crawl(
-                    const {
-                        &SyntaxSet::new(&[
-                            SyntaxKind::Identifier,
-                            SyntaxKind::NakedIdentifier,
-                            SyntaxKind::QuotedIdentifier,
-                            SyntaxKind::Literal,
-                            SyntaxKind::Dash,
-                            SyntaxKind::Dot,
-                            SyntaxKind::Star,
-                        ])
-                    },
-                    true,
-                    &SyntaxSet::EMPTY,
-                    true,
-                ) {
-                    if !elem.is_type(SyntaxKind::Dot) {
-                        if elem.is_type(SyntaxKind::Identifier)
-                            || elem.is_type(SyntaxKind::NakedIdentifier)
-                            || elem.is_type(SyntaxKind::QuotedIdentifier)
-                        {
-                            let raw = elem.raw();
-                            let elem_raw = raw.trim_matches('`');
-                            let elem_subparts = elem_raw.split(".").collect_vec();
-                            let elem_subparts_count = elem_subparts.len();
-
-                            for (idx, part) in enumerate(elem_subparts) {
-                                parts.push(part.to_smolstr());
-                                elems_for_parts.push(elem.clone());
-
-                                if idx != elem_subparts_count - 1 {
-                                    flush(&mut parts, &mut elems_for_parts);
-                                }
-                            }
-                        } else {
-                            parts.push(elem.raw().to_smolstr());
-                            elems_for_parts.push(elem);
-                        }
-                    } else {
-                        flush(&mut parts, &mut elems_for_parts);
-                    }
-                }
-
-                if !parts.is_empty() {
-                    flush(&mut parts, &mut elems_for_parts);
-                }
-
-                acc
-            }
-            ObjectReferenceKind::Object | ObjectReferenceKind::Table => {
-                let mut acc = Vec::new();
-
-                for elem in self.0.recursive_crawl(
-                    const {
-                        &SyntaxSet::new(&[
-                            SyntaxKind::Identifier,
-                            SyntaxKind::NakedIdentifier,
-                            SyntaxKind::QuotedIdentifier,
-                        ])
-                    },
-                    true,
-                    &SyntaxSet::EMPTY,
-                    true,
-                ) {
-                    acc.extend(self.iter_reference_parts(elem));
-                }
-
-                acc
-            }
-            ObjectReferenceKind::WildcardIdentifier => {
-                let mut acc = Vec::new();
-
-                for elem in self.0.recursive_crawl(
-                    const {
-                        &SyntaxSet::new(&[
-                            SyntaxKind::Identifier,
-                            SyntaxKind::Star,
-                            SyntaxKind::NakedIdentifier,
-                            SyntaxKind::QuotedIdentifier,
-                        ])
-                    },
-                    true,
-                    &SyntaxSet::EMPTY,
-                    true,
-                ) {
-                    acc.extend(self.iter_reference_parts(elem));
-                }
-
-                acc
-            }
-        }
-    }
-
-    fn iter_reference_parts(&self, elem: ErasedSegment) -> Vec<ObjectReferencePart> {
-        let mut acc = Vec::new();
-
-        let raw = elem.raw();
-        let parts = raw.split('.');
-
-        for part in parts {
-            acc.push(ObjectReferencePart { part: part.into(), segments: vec![elem.clone()] });
-        }
-
-        acc
-    }
-}
-
-pub struct JoinClauseSegment(pub ErasedSegment);
-
-impl JoinClauseSegment {
-    pub fn eventual_aliases(&self) -> Vec<(ErasedSegment, AliasInfo)> {
-        let mut buff = Vec::new();
-
-        let from_expression =
-            self.0.child(const { &SyntaxSet::new(&[SyntaxKind::FromExpressionElement]) }).unwrap();
-        let alias = FromExpressionElementSegment(from_expression.clone()).eventual_alias();
-
-        buff.push((from_expression.clone(), alias));
-
-        for join_clause in self.0.recursive_crawl(
-            const { &SyntaxSet::new(&[SyntaxKind::JoinClause]) },
-            true,
-            const { &SyntaxSet::single(SyntaxKind::SelectStatement) },
-            true,
-        ) {
-            if join_clause.id() == join_clause.id() {
-                continue;
-            }
-
-            let aliases = JoinClauseSegment(join_clause).eventual_aliases();
-
-            if !aliases.is_empty() {
-                buff.extend(aliases);
-            }
-        }
-
-        buff
-    }
 }
 
 #[cfg(test)]
