+++ crates/lib/src/rules/aliasing/al04.rs
@@ -5,10 +5,10 @@ use smol_str::SmolStr;
 
 use crate::core::config::Value;
 use crate::core::dialects::common::{AliasInfo, ColumnAliasInfo};
+use crate::core::parser::segments::object_reference::ObjectReferenceSegment;
 use crate::core::rules::base::{Erased, ErasedRule, LintResult, Rule, RuleGroups};
 use crate::core::rules::context::RuleContext;
 use crate::core::rules::crawlers::{Crawler, SegmentSeekerCrawler};
-use crate::dialects::ansi::ObjectReferenceSegment;
 use crate::dialects::{SyntaxKind, SyntaxSet};
 use crate::helpers::IndexSet;
 use crate::utils::analysis::select::get_select_statement_info;
