+++ crates/lib/src/rules/aliasing/al05.rs
@@ -8,10 +8,10 @@ use crate::core::dialects::base::Dialect;
 use crate::core::dialects::common::AliasInfo;
 use crate::core::dialects::init::DialectKind;
 use crate::core::parser::segments::base::ErasedSegment;
+use crate::core::parser::segments::object_reference::ObjectReferenceLevel;
 use crate::core::rules::base::{Erased, ErasedRule, LintFix, LintResult, Rule, RuleGroups};
 use crate::core::rules::context::RuleContext;
 use crate::core::rules::crawlers::{Crawler, SegmentSeekerCrawler};
-use crate::dialects::ansi::ObjectReferenceLevel;
 use crate::dialects::{SyntaxKind, SyntaxSet};
 use crate::utils::analysis::query::Query;
 use crate::utils::analysis::select::get_select_statement_info;
