+++ crates/lib/src/rules/references/rf03.rs
@@ -8,10 +8,10 @@ use crate::core::config::Value;
 use crate::core::dialects::common::{AliasInfo, ColumnAliasInfo};
 use crate::core::dialects::init::DialectKind;
 use crate::core::parser::segments::base::{ErasedSegment, SegmentBuilder, Tables};
+use crate::core::parser::segments::object_reference::ObjectReferenceSegment;
 use crate::core::rules::base::{Erased, ErasedRule, LintFix, LintResult, Rule, RuleGroups};
 use crate::core::rules::context::RuleContext;
 use crate::core::rules::crawlers::{Crawler, SegmentSeekerCrawler};
-use crate::dialects::ansi::ObjectReferenceSegment;
 use crate::dialects::{SyntaxKind, SyntaxSet};
 use crate::helpers::capitalize;
 use crate::utils::analysis::query::Query;
